restructured model and query mocks and stubs into their own modules
