+++ index.js
@@ -1,11 +1,11 @@
 'use strict';
 
 const model = {};
-model.mock = require('./lib/model/model-mock');
-model.stub = require('./lib/model/model-stub');
+model.mock = require('./lib/model-mock');
+model.stub = require('./lib/model-stub');
 module.exports.model = model;
 
 const query = {};
-query.mock = require('./lib/query/query-mock');
-query.stub = require('./lib/query/query-stub');
+query.mock = require('./lib/query-mock');
+query.stub = require('./lib/query-stub');
 module.exports.query = query;
