+++ lib/function-mappings/model-query-functions.js
