+++ lib/function-mappings/query-chainable-functions.js
