+++ lib/model-mock/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const modelQueryFunctions = require('./model-query-functions');
-const queryMock = require('../query/query-mock');
+const modelQueryFunctions = require('../function-mappings/model-query-functions');
+const queryMock = require('../query-mock');
 const deepEqual = require('deep-equal');
 const callbackUtil = require('../util/callback-util');
