+++ lib/model-mock/index.test.js
@@ -1,10 +1,9 @@
 'use strict';
 
-const chai = require('chai');
-const expect = chai.expect;
-const modelMock = require('./model-mock');
-const modelQueryFunctions = require('./model-query-functions');
-const chainableQueryFunctions = require('../query/chainable-query-functions');
+const expect = require('chai').expect;
+const modelMock = require('./index');
+const modelQueryFunctions = require('../function-mappings/model-query-functions');
+const queryChainableFunctions = require('../function-mappings/query-chainable-functions');
 
 describe('model-mock', () => {
   describe('mock()', () => {
@@ -67,7 +66,7 @@ describe('model-mock', () => {
   for (const modelQueryFunction of modelQueryFunctions) {
     describe(`${modelQueryFunction.name}()`, () => {
       // Determine if the function supports callback
-      const supportsCallback = chainableQueryFunctions.find(e => e.name === modelQueryFunction.queryName).callback;
+      const supportsCallback = queryChainableFunctions.find(e => e.name === modelQueryFunction.queryName).callback;
 
       it('throws an error when called with arguments that the tester did not specify', (done) => {
         // Setup our mocks
