+++ lib/model-stub/index.js
@@ -1,7 +1,7 @@
 'use-strict';
 
-const modelQueryFunctions = require('./model-query-functions');
-const queryStub = require('../query/query-stub');
+const modelQueryFunctions = require('../function-mappings/model-query-functions');
+const queryStub = require('../query-stub');
 const callbackUtil = require('../util/callback-util');
 
 module.exports = (modelName, schema) => {
