+++ lib/model-stub/index.test.js
@@ -1,10 +1,9 @@
 'use strict';
 
-const chai = require('chai');
-const expect = chai.expect;
-const modelStub = require('./model-stub');
-const modelQueryFunctions = require('./model-query-functions');
-const chainableQueryFunctions = require('../query/chainable-query-functions');
+const expect = require('chai').expect;
+const modelStub = require('./index');
+const modelQueryFunctions = require('../function-mappings/model-query-functions');
+const queryChainableFunctions = require('../function-mappings/query-chainable-functions');
 
 describe('model-stub', () => {
   describe('stub()', () => {
@@ -67,7 +66,7 @@ describe('model-stub', () => {
   for (const modelQueryFunction of modelQueryFunctions) {
     describe(`${modelQueryFunction.name}()`, () => {
       // Determine if the function supports callback
-      const supportsCallback = chainableQueryFunctions.find(e => e.name === modelQueryFunction.queryName).callback;
+      const supportsCallback = queryChainableFunctions.find(e => e.name === modelQueryFunction.queryName).callback;
 
       it('does not exist when called without defined behavior', (done) => {
         // Create a Mongoose Model stub
