+++ lib/query-mock/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const callbackUtil = require('../util/callback-util');
-const chainableQueryFunctions = require('./chainable-query-functions');
+const queryChainableFunctions = require('../function-mappings/query-chainable-functions');
 const deepEqual = require('deep-equal');
 
 // Unfortunately can't just return the Query mock directly due to caching so instead we return a builder function
@@ -57,7 +57,7 @@ module.exports = () => {
   Query.proto.proto = Query.proto;
 
   // Set up the chainable query functions
-  for (const chainableFunction of chainableQueryFunctions) {
+  for (const chainableFunction of queryChainableFunctions) {
     Query.prototype[chainableFunction.name] = function (...args) {
       const argsWithoutCallback = callbackUtil.getArgumentsWithoutCallback(args);
       this.executionChain.chain.push({ name: chainableFunction.name, args: argsWithoutCallback });
