+++ lib/query-mock/index.test.js
@@ -1,14 +1,13 @@
 'use strict';
 
-const chai = require('chai');
-const expect = chai.expect;
-const queryMock = require('./query-mock');
-const chainableQueryFunctions = require('./chainable-query-functions');
+const expect = require('chai').expect;
+const queryMock = require('./index');
+const queryChainableFunctions = require('../function-mappings/query-chainable-functions');
 
 describe('query-mock', () => {
 
   // Generate tests for query functions
-  for (const chainableFunction of chainableQueryFunctions) {
+  for (const chainableFunction of queryChainableFunctions) {
     describe(`${chainableFunction.name}.withArgs()`, () => {
       it('allows assertion chaining', (done) => {
         const Query = queryMock();
