+++ lib/query-stub/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const chainableQueryFunctions = require('./chainable-query-functions');
+const queryChainableFunctions = require('../function-mappings/query-chainable-functions');
 const callbackUtil = require('../util/callback-util');
 
 // Unfortunately can't just return the Query stub directly due to caching so instead we return a builder function
@@ -46,7 +46,7 @@ module.exports = () => {
   };
 
   // Set up the chainable query functions
-  for (const chainableFunction of chainableQueryFunctions) {
+  for (const chainableFunction of queryChainableFunctions) {
     if (chainableFunction.callback === true) {
       Query.prototype[chainableFunction.name] = function(...args) {
         const callback = callbackUtil.getCallback(args);
