+++ lib/query-stub/index.test.js
@@ -1,14 +1,13 @@
 'use strict';
 
-const chai = require('chai');
-const expect = chai.expect;
-const queryStub = require('./query-stub');
-const chainableQueryFunctions = require('./chainable-query-functions');
+const expect = require('chai').expect;
+const queryStub = require('./index');
+const queryChainableFunctions = require('../function-mappings/query-chainable-functions');
 
 describe('query-stub', () => {
 
   // Generate tests for query functions
-  for (const chainableFunction of chainableQueryFunctions) {
+  for (const chainableFunction of queryChainableFunctions) {
     describe(`returns() ${chainableFunction.name}`, () => {
 
       // Don't test callback capability for functions that do not allow a callback to be passed
